Move download() method from api.py to download.py
